fix bug when calculating intersection of areas in gis module
The current intersection of areas is using tree.query(o) but this provides the index of the elements in "tree" that overlaps with "o". This gives an error because it attempts to calculate the intersection of an index "o" with a geometry "d". The intersection must be calculated between the geometry corresponding to the index "o", that is "orig[o]" and the element in the tree "d"